# Faster sampling from `Gumbel`