--- a/Project.toml
+++ b/Project.toml
@@ -1,7 +1,7 @@
 name = "Distributions"
 uuid = "31c24e10-a181-5473-b8eb-7969acd0382f"
 authors = ["JuliaStats"]
-version = "0.24.16"
+version = "0.24.17"
 
 [deps]
 FillArrays = "1a297f60-69ca-5386-bcde-b61e274b549b"