--- a/test/runtests.jl
+++ b/test/runtests.jl
@@ -61,6 +61,7 @@ const tests = [
     "soliton",
     "skewnormal",
     "chi",
+    "gumbel",
 ]
 
 printstyled("Running tests:\n", color=:blue)