On master:
```julia
julia> using Distributions, BenchmarkTools

julia> @btime rand($(Gumbel()));
  34.065 ns (0 allocations: 0 bytes)

julia> @btime rand($(Gumbel()), 100);
  3.433 μs (1 allocation: 896 bytes)
```
With this PR:
```julia
julia> using Distributions, BenchmarkTools

julia> @btime rand($(Gumbel()));
  22.429 ns (0 allocations: 0 bytes)

julia> @btime rand($(Gumbel()), 100);
  1.655 μs (1 allocation: 896 bytes)
```